refactor: move stake pool types out of the Cardano.Wallet.Primitive.Types into appropriate dedicated modules;
